Declarations refactor (types.ts / globals.d.ts)
typings.d.ts has been refactored—utility/data types are now in types.ts, and globals/module augmentations in globals.d.ts. Publicly consumed types are now included in declaration emit, module augmentations skipped. Missing type import and export have also been added.
