+++ src/globals.d.ts
@@ -1,0 +1,6 @@
+declare module '*.woff';
+
+declare module '*.css' {
+  const content: { [className: string]: string }
+  export default content
+}
