--- src/woff.d.ts
@@ -1,1 +1,0 @@
-declare module '*.woff';
